requirements.txt
deeppavlov==1.3.0
dff==0.4.2
fastapi==0.103.1
fire==0.5.0
Flask==1.1.2
llama_cpp_python==0.1.84
playsound==1.3.0
PyAudio==0.2.13
pydantic==1.10.9
pydub==0.25.1
PyQt5==5.15.9
pyTelegramBotAPI==4.13.0
pyttsx3==2.90
requests==2.28.0
sounddevice==0.4.6
SQLAlchemy==2.0.15
tzlocal==5.0.1
uvicorn==0.23.2
vosk==0.3.45
